refactor: added base store external profile action
