+++ app/Actions/Models/List/ExternalProfile/ExternalEntry/Site/AnilistExternalEntryAction.php
@@ -52,7 +52,7 @@ class AnilistExternalEntryAction extends BaseExternalEntryAction
     /**
      * Get the id of the external user.
      *
-     * @return string
+     * @return int|null
      */
     public function getId(): ?int
     {
